trying to change order of boxes -
I got something to happen when I click on a box, but it's not what I want. Also, I had to edit the html to include a meta charset = utf-8 because my browser was freaking out without it. I don't think I changed anything substantive with the css - I was just trying to confirm which element to change in order to move the boxes. I think I put it back the way it was to begin with.