feat: Adds new title animation in index.js
Stretch goal for Fun Bus site is complete. Title is now broken into multiple spans and each letter comes in from a different direction and with different rotations. At the end of the run, each letter has an elastic bounce. At the end of that animation, the title, now together, bows slightly.