style: Change to title flyin to account for events
The title fly-in animations were loading automatically and I wanted them to be loaded via a page load event in case I wanted to change it to a click or mouseover or similar event. An .addEventListener has been added and the TweenMax methods are now loaded from a function on page load.