Complete first stretch task
I used GSAP to handle animations.  Honestly it ended up being more difficult that just using basic CSS to do them.  There were weird interactions that limited the animations and there was no way to target ::before pseudo-classes which I needed in this case. Their docs said the one of the plugins would allow that behavior, but after a bunch of testing, I was never able to get it to work. I'll spend more time with it in the future, but for right now it makes things more difficult.