Create Mouseenter Event and Load Event
Created a fun mouseenter even that changes each character in the main h2 element a differnt color then back to black after a bit.
Also created a load event that prompts users to enter their name and takes that input and adds it to the beginning of the main h2 elememt.