feat: Add newletter sign up pop-up
In this commit, using the javascript 'load' event I added an event handler that listens for
anytime the web page gets reloaded. Whenever the page gets reloaded, the visibility of the
signup section of the html document is change from hidden(which happens to be its default)
to visible. This makes it possible for a user to get reminded to sign-up for the newsletter